+++ Editor/FriggPropertyDrawer.cs
@@ -16,7 +16,7 @@
         public static FriggPropertyDrawer GetCustomDrawer(FriggProperty property) {
             var meta = property.MetaInfo.MemberInfo;
 
-            if (meta.IsDefined(typeof(ReadonlyAttribute))) {
+            if (!CoreUtilities.IsWritable(property.MetaInfo.MemberInfo)) {
                 return new ReadonlyPropertyDrawer(property);
             }
